+++ src/main/java/dev/wroud/mc/worlds/command/CreateCommand.java
@@ -2,6 +2,7 @@ package dev.wroud.mc.worlds.command;
 
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
+import com.mojang.brigadier.exceptions.DynamicCommandExceptionType;
 import com.mojang.brigadier.exceptions.SimpleCommandExceptionType;
 import com.mojang.logging.LogUtils;
 
@@ -9,57 +10,90 @@ import dev.wroud.mc.worlds.McWorldInitializer;
 import dev.wroud.mc.worlds.manadger.LevelData;
 import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.commands.arguments.ResourceKeyArgument;
 import net.minecraft.commands.arguments.ResourceLocationArgument;
 
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.LongArgumentType;
+
 import net.minecraft.core.registries.Registries;
+import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.dimension.BuiltinDimensionTypes;
 import net.minecraft.world.level.dimension.LevelStem;
+import net.minecraft.world.level.levelgen.WorldOptions;
 
 import static me.drex.message.api.LocalizedMessage.builder;
 import static me.drex.message.api.LocalizedMessage.localized;
 import static net.minecraft.commands.Commands.argument;
 import static net.minecraft.commands.Commands.literal;
 
+import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 
 public class CreateCommand {
 	private static final Logger LOGGER = LogUtils.getLogger();
+	private static final DynamicCommandExceptionType ERROR_INVALID_LEVEL_STEM = new DynamicCommandExceptionType(
+			object -> Component.translatableEscape("advancement.advancementNotFound", object));
 
 	public static LiteralArgumentBuilder<CommandSourceStack> build() {
 		return literal("create")
 				.requires(Permissions.require("dev.wroud.mc.worlds.command.create", 2))
 				.then(
 						argument("id", ResourceLocationArgument.id())
-								.executes(context -> createWorld(context.getSource(), ResourceLocationArgument.getId(context, "id"), 0))
+								.executes(
+										context -> createWorld(context.getSource(), ResourceLocationArgument.getId(context, "id"), null,
+												null))
+								.then(
+										argument("type", ResourceKeyArgument.key(Registries.LEVEL_STEM))
+												.executes(
+														context -> createWorld(
+																context.getSource(),
+																ResourceLocationArgument.getId(context, "id"),
+																ResourceKeyArgument.getRegistryKey(context, "type", Registries.LEVEL_STEM,
+																		ERROR_INVALID_LEVEL_STEM),
+																null))
+												.then(
+														argument("seed", LongArgumentType.longArg())
+																.executes(
+																		context -> createWorld(context.getSource(),
+																				ResourceLocationArgument.getId(context, "id"),
+																				ResourceKeyArgument.getRegistryKey(context, "type", Registries.LEVEL_STEM,
+																						ERROR_INVALID_LEVEL_STEM),
+																				LongArgumentType.getLong(context, "seed")))))
 								.then(
 										argument("seed", LongArgumentType.longArg())
 												.executes(
 														context -> createWorld(context.getSource(), ResourceLocationArgument.getId(context, "id"),
+																null,
 																LongArgumentType.getLong(context, "seed")))));
 	}
 
-	public static int createWorld(CommandSourceStack source, ResourceLocation id, long seed)
+	public static int createWorld(CommandSourceStack source, ResourceLocation id,
+			@Nullable ResourceKey<LevelStem> type, @Nullable Long seed)
 			throws CommandSyntaxException {
 
-		LOGGER.info("Creating new world with id: {}, seed: {}", id, seed);
+		if (type == null) {
+			type = LevelStem.OVERWORLD;
+		}
+
+		if (seed == null) {
+			seed = WorldOptions.randomSeed();
+		}
+
+		LOGGER.info("Creating new world with id: {}, seed: {}, type: {}", id, seed, type);
 		var server = source.getServer();
 		validLevelId(id, server);
 
 		var registry = server.registryAccess();
-		var dimensionType = registry
-				.lookupOrThrow(Registries.DIMENSION_TYPE);
-		var type = dimensionType.getOrThrow(BuiltinDimensionTypes.OVERWORLD);
-		var levelStem = new LevelStem(type,
-				server.overworld().getChunkSource().getGenerator());
-		LOGGER.info("Using dimension type: {}", type);
-		var levelData = LevelData.overworldDefault(id, levelStem, seed);
+		var levelStemRegistry = registry.lookupOrThrow(Registries.LEVEL_STEM);
+
+		LevelStem levelStem = levelStemRegistry.getValue(type);
+		LOGGER.info("Using dimension type: {}", levelStem.type());
+		var levelData = LevelData.getDefault(id, levelStem, seed, true);
 
 		var handle = McWorldInitializer.getMcWorld(server).loadOrCreate(id,
 				levelData);
@@ -77,7 +111,8 @@ public class CreateCommand {
 		ResourceKey<Level> resourceKey = ResourceKey.create(Registries.DIMENSION, id);
 		ServerLevel level = server.getLevel(resourceKey);
 		if (level != null) {
-			throw new SimpleCommandExceptionType(localized("dev.wroud.mc.worlds.command.create.exception.world_already_exists"))
+			throw new SimpleCommandExceptionType(
+					localized("dev.wroud.mc.worlds.command.create.exception.world_already_exists"))
 					.create();
 		}
 	}