+++ src/main/java/dev/wroud/mc/worlds/mixin/MinecraftServerAccessor.java
@@ -5,6 +5,7 @@ import java.util.concurrent.Executor;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
+import org.spongepowered.asm.mixin.gen.Invoker;
 
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.MinecraftServer;
@@ -12,6 +13,8 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.progress.ChunkProgressListenerFactory;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.storage.LevelStorageSource;
+import net.minecraft.world.level.storage.ServerLevelData;
+import net.minecraft.world.level.storage.WorldData;
 
 @Mixin(MinecraftServer.class)
 public interface MinecraftServerAccessor {
@@ -26,4 +29,14 @@ public interface MinecraftServerAccessor {
 
     @Accessor("levels")
     Map<ResourceKey<Level>, ServerLevel> getLevels();
+
+    @Invoker("setupDebugLevel")
+    void invokeSetupDebugLevel(WorldData worldData);
+
+    @Invoker("setInitialSpawn")
+    static void invokeSetInitialSpawn(ServerLevel serverLevel, ServerLevelData serverLevelData,
+            boolean generateBonusChest,
+            boolean debug) {
+        throw new AssertionError("Mixin invoker should not be called directly");
+    }
 }