+++ src/main/java/dev/wroud/mc/worlds/util/DimensionDetectionUtil.java
@@ -3,6 +3,8 @@ package dev.wroud.mc.worlds.util;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.dimension.BuiltinDimensionTypes;
+import net.minecraft.world.level.dimension.DimensionType;
+
 import org.jetbrains.annotations.Nullable;
 
 public class DimensionDetectionUtil {
@@ -19,19 +21,29 @@ public class DimensionDetectionUtil {
         return level.dimensionTypeRegistration().is(BuiltinDimensionTypes.OVERWORLD);
     }
 
-    public static @Nullable ResourceKey<Level> getVanillaDimensionMapping(Level level) {
-        if (isEndLikeDimension(level)) {
+    public static ResourceKey<Level> getVanillaDimensionMapping(Level level) {
+        var mapping = getVanillaDimensionMapping(level.dimensionTypeRegistration().unwrapKey().orElse(null));
+
+        if (mapping == null) {
+            return level.dimension();
+        }
+
+        return mapping;
+    }
+
+    public static @Nullable ResourceKey<Level> getVanillaDimensionMapping(ResourceKey<DimensionType> type) {
+        if (type.equals(BuiltinDimensionTypes.END)) {
             return Level.END;
         }
-        
-        if (isNetherLikeDimension(level)) {
+
+        if (type.equals(BuiltinDimensionTypes.NETHER)) {
             return Level.NETHER;
         }
-        
-        if (isOverworldLikeDimension(level)) {
+
+        if (type.equals(BuiltinDimensionTypes.OVERWORLD)) {
             return Level.OVERWORLD;
         }
-        
+
         return null;
     }