+++ CHANGELOG.md
@@ -14,6 +14,17 @@
 
 ### Security
 
+## 1.1.0 - 2025-09-04
+
+### Added
+
+- Support for Nether and End dimensions creation
+- Proper Overworld initialization and spawn setting
+
+### Fixed
+
+- End world teleportation
+
 ## 1.0.1 - 2025-09-04
 
 ### Added